moving measure/reset extensions from QuantumCircuit to their own files
