+++ qiskit/_measure.py
@@ -9,6 +9,7 @@
 Quantum measurement in the computational basis.
 """
 from ._instruction import Instruction
+from ._quantumcircuit import QuantumCircuit
 
 
 class Measure(Instruction):
@@ -30,3 +31,29 @@ class Measure(Instruction):
     def reapply(self, circuit):
         """Reapply this gate to corresponding qubits."""
         self._modifiers(circuit.measure(self.arg[0], self.arg[1]))
+
+
+def measure(self, qubit, cbit):
+    """Measure quantum bit into classical bit (tuples).
+
+    Returns:
+        qiskit.Gate: the attached measure gate.
+
+    Raises:
+        QISKitError: if qubit is not in this circuit or bad format;
+            if cbit is not in this circuit or not creg.
+    """
+    if isinstance(qubit, QuantumRegister) and \
+       isinstance(cbit, ClassicalRegister) and len(qubit) == len(cbit):
+        instructions = InstructionSet()
+        for i in range(qubit.size):
+            instructions.add(self.measure((qubit, i), (cbit, i)))
+        return instructions
+
+    self._check_qubit(qubit)
+    self._check_creg(cbit[0])
+    cbit[0].check_range(cbit[1])
+    return self._attach(Measure(qubit, cbit, self))
+
+
+QuantumCircuit.measure = measure
